src/DeltaShell.Plugins.NetworkEditor.Gui/Forms/PropertyGrid/PropertyInfoCreation/HydroNodePropertyInfoCreationContext.cs
using DelftTools.Hydro;
using DelftTools.Utils.Guards;
using DelftTools.Utils.Validation.NameValidation;
using DeltaShell.NGHS.Common.Gui;
using DeltaShell.NGHS.Common.Gui.PropertyGrid.PropertyInfoCreation;

namespace DeltaShell.Plugins.NetworkEditor.Gui.Forms.PropertyGrid.PropertyInfoCreation
{
    /// <summary>
    /// Provides the creation context for a <see cref="PropertyInfoCreator"/> that should create the
    /// <see cref="DelftTools.Shell.Gui.PropertyInfo"/> for <see cref="HydroNode"/> data.
    /// </summary>
    public sealed class HydroNodePropertyInfoCreationContext : IPropertyInfoCreationContext<HydroNode, HydroNodeProperties>
    {
        /// <inheritdoc/>
        public void CustomizeProperties(HydroNodeProperties properties, GuiContainer guiContainer)
        {
            Ensure.NotNull(properties, nameof(properties));
            Ensure.NotNull(guiContainer, nameof(guiContainer));

            properties.NameValidator.AddValidator(new UniqueNameValidator(properties.Data.HydroNetwork.HydroNodes));
        }
    }
}